requirements.txt
pytest==7.1.2
cffi==1.14.4
cryptography==3.3.2
pycparser==2.20
PyMySQL==0.10.1
python-dotenv==0.15.0
six==1.15.0
SQLAlchemy==1.4.39
pandas==1.1.4
